Move: move domain to service package
